Refactor toast
